% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_new.R
\name{install_new}
\alias{install_new}
\title{Install a new package}
\usage{
install_new(pkg, repo = "http://cran.revolutionanalytics.com/",
  libPath = .libPaths()[1])
}
\arguments{
\item{pkg}{String. The package to install}

\item{repo}{Sring, Repository. Default: 'http://cran.revolutionanalytics.com/'}

\item{libPath}{String Ex:.libPaths()[1]}
}
\description{
Install a new package
}
\examples{
library(lwPackageHelperR)
install_new("scales")
}
